rag-book-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-google-genai==0.0.8
langchain-qdrant==0.1.3
qdrant-client==1.9.0
chromadb==0.4.22
PyPDF2==3.0.1
pypdf==3.17.1
python-dotenv==1.0.0
tiktoken==0.5.1
sentence-transformers==2.2.2
torch>=2.1.1